(WIP) Turborepo working, first core elemnts of protocol.
At the very base of the protocol we have a tree of problems being
analyzed into solveable components. So I wrote some types to represent
such a thing and began writing a super simple mock prompt and function.

Next thing to do is to finish the mocks and write a simple execution
engine "mock".