(WIP) Continue with the process control direction
I probably spent too much time thinking about the "cybernetic"
architecture on this one. That said, I do actually think that in this
very specific project, the architecture is extremely important because
the architecture is kind of the whole point. But I know I won't get the
architecture right if I don't do it iteratively. Sooo, I made a few
conclusions that I think will be kind of future proof and let me
continue with the implementation.

It's hard because I have this very strong feeling that the golden
solution is rightttt over the horizon, right on the tip of my tongue.

One of the difficulties is that I'm mixing a few different "metaphors"
for the architecture. On the one hand I'm writing a programming
language but on the other hand it's a kind of cybernetic machine. I'm a
little bit worried that if I go too far into the implementation I'm going
to lose sight of the correct solution, I'm going to get caught up in a
bunch of incidental details.

The next thing I'm working on and thinking through is the implementation
of the engine's loop. Should it be an event sourced system, a functional
one, or an object-oriented one? Or what extent should it be some of
each?