WIP Nodes, memory, source, sink
I'm thinking through what exactly the network primitives are. Do we want
to just link up a bunch of memory blocks? Do we want sources and sinks?
Do we want a 2-layer system where there's inboxes and memory? Is a
Neuron an abstraction or a primitive? These are some of the things
running through my head at the moment.