WIP Sestina network
The idea is to create a network that can successfully generate a perfect
sestina. This is a great candidate for number one because gpt-3.5-turbo
gets pretty close but fails in an analytically-detectable way. So
basically I need to write a bunch of analytic functions that test
different properties of a sestina. How close to 100% reliability can I
get?