Feedback loop v0.0.0
In Sestina generation an incredibly naive approach to feedback seems to
help gpt-3.5-turbo "converge" to a correct answer in roughly 1/3 the
attempts. I'm positive a less than naive approach could do significantly
better.

One thing I'm thinking about is this idea of genuine promptlessness. How
can we design the feedback mechanism so that it's just language models
talking to each other? Do we want promptlessness or do we want BYOP? Or
something in between?